Antes de limpiar Generators

+++ JDBC/DAO/src/main/java/dam/ad/personas/db/hsqldb/DbPersonaDAO.java
@@ -1,8 +1,8 @@
 package dam.ad.personas.db.hsqldb;
 
 import dam.ad.dao.jdbc.DbDAO;
-import dam.ad.personas.model.Persona;
-import dam.ad.personas.model.Sexo;
+import dam.ad.model.personas.Persona;
+import dam.ad.model.personas.Sexo;
 
 import javax.sql.DataSource;
 import java.sql.PreparedStatement;
@@ -38,7 +38,7 @@ public class DbPersonaDAO extends DbDAO<Persona> {
                 resultSet.getString("apellidos"),
                 Sexo.fromInicial(resultSet.getString("sexo")),
                 resultSet.getObject("nacimiento", LocalDate.class),
-                resultSet.getDouble("ingresos")
+                resultSet.getFloat("ingresos")
         );
     }
